One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/common/addstore/addstore.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { AddstoreComponent } from './addstore.component';

describe('Component: Addstore', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [AddstoreComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([AddstoreComponent],
      (component: AddstoreComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(AddstoreComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(AddstoreComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-addstore></app-addstore>
  `,
  directives: [AddstoreComponent]
})
class AddstoreComponentTestController {
}

